refactor: colocate persistence unit tests
